init game entity

+++ mini-boardgame-server/mini-boardgame-server-app/src/main/resources/sql/schema/security_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users(
-  id BIGINT PRIMARY KEY AUTO_INCREMENT,
+  id int PRIMARY KEY AUTO_INCREMENT,
   `username` VARCHAR(511) NOT NULL,
   `password` VARCHAR(511) NOT NULL,
   enabled BOOLEAN NOT NULL,
@@ -15,18 +15,18 @@ CREATE TABLE IF NOT EXISTS authorities (
 
 /*
 CREATE TABLE IF NOT EXISTS groups (
-  id BIGINT PRIMARY KEY AUTO_INCREMENT,
+  id int PRIMARY KEY AUTO_INCREMENT,
   group_name VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS group_authorities (
-  group_id BIGINT NOT NULL,
+  group_id int NOT NULL,
   authority VARCHAR(50) NOT NULL,
   CONSTRAINT fk_group_authorities_group FOREIGN KEY(group_id) REFERENCES groups(id)
 );
 
 CREATE TABLE IF NOT EXISTS group_members (
-  id BIGINT PRIMARY KEY AUTO_INCREMENT,
+  id int PRIMARY KEY AUTO_INCREMENT,
   username VARCHAR(512) NOT NULL,
   group_id BIGINT NOT NULL,
   CONSTRAINT fk_group_members_users FOREIGN KEY(username) REFERENCES users(username),
